requirements.txt
asgiref==3.8.1
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.0
pytest-django==4.8.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
setuptools==72.2.0
sqlparse==0.5.0
uritemplate==4.1.1
